first strategy for enemy - move towards hero with speed 2

+++ Scripts/GeroiController.cs
@@ -3,11 +3,11 @@ using System;
 
 public partial class GeroiController : Controller
 {
-    public GeroiController(int x0, int y0, int xMax, int yMax) : base(x0, y0, xMax, yMax)
+    public GeroiController(float x0, float y0, float xMax, float yMax) : base(x0, y0, xMax, yMax)
     {
     }
 
-    public override Tuple<float, int> Decision(int x, int y, float angle, int otherX, int otherY, int health)
+    public override Tuple<float, int> Decision(float x, float y, float angle, float otherX, float otherY, int health)
     {
         return new Tuple<float, int>(0.1f, 0);
     }
